+++ src/main/java/org/cicirello/math/rand/RandomIndexer.java
@@ -662,8 +662,68 @@ public final class RandomIndexer {
     final int z2 = z1 << 1;
     int i = nextInt(z2 + window - 1, gen);
     int j = nextInt(window, gen);
-    setAndAdjustWindowedPair(result, i, j, z1, z2);
+    if (i < z2) {
+      int x = i & 1;
+      result[x] = i >> 1;
+      result[x ^ 1] = result[x] + 1 + j;
+    } else {
+      i -= z1;
+      j += z1;
+      result[0] = i == j ? n - 1 : i;
+      result[1] = j;
+    }
     return result;
+  }
+
+  /**
+   * Generates a random sample of 2 integers, i, j, without replacement, from the set of integers in
+   * the interval [0, n), such that |i-j| &le; window. All pairs that satisfy the window constraint
+   * are equally likely.
+   *
+   * <p>The runtime is O(1).
+   *
+   * <p>This method uses ThreadLocalRandom as the pseudorandom number generator, and is thus safe,
+   * and efficient (i.e., non-blocking), for use with threads.
+   *
+   * @param n The number of integers to choose from.
+   * @param window The maximum difference between the integers of the pair.
+   * @return A pair of randomly chosen integers, i, j, from the interval [0, n), such that |i-j|
+   *     &le; window.
+   * @throws IllegalArgumentException if window &lt; 1 or n &lt; 2.
+   */
+  public static IndexPair nextWindowedIntPair(int n, int window) {
+    return nextWindowedIntPair(n, window, ThreadLocalRandom.current());
+  }
+
+  /**
+   * Generates a random sample of 2 integers, i, j, without replacement, from the set of integers in
+   * the interval [0, n), such that |i-j| &le; window. All pairs that satisfy the window constraint
+   * are equally likely.
+   *
+   * <p>The runtime is O(1).
+   *
+   * @param n The number of integers to choose from.
+   * @param window The maximum difference between the integers of the pair.
+   * @param gen Source of randomness.
+   * @return A pair of randomly chosen integers, i, j, from the interval [0, n), such that |i-j|
+   *     &le; window.
+   * @throws IllegalArgumentException if window &lt; 1 or n &lt; 2.
+   */
+  public static IndexPair nextWindowedIntPair(int n, int window, RandomGenerator gen) {
+    if (window >= n - 1) return nextIntPair(n, gen);
+    final int z1 = n - window;
+    final int z2 = z1 << 1;
+    int i = nextInt(z2 + window - 1, gen);
+    int j = nextInt(window, gen);
+    if (i < z2) {
+      final int rightBit = i & 1;
+      i >>= 1;
+      return rightBit == 0 ? new IndexPair(i, i + 1 + j) : new IndexPair(i + 1 + j, i);
+    } else {
+      i -= z1;
+      j += z1;
+      return new IndexPair(i == j ? n - 1 : i, j);
+    }
   }
 
   /**
@@ -731,7 +791,7 @@ public final class RandomIndexer {
     if (window >= n - 1) return nextIntTriple(n, result, gen);
     result = ArrayMinimumLengthEnforcer.enforce(result, 3);
     final int z1 = n - window;
-    final int z3 = (z1 << 1) + z1;
+    final int z3 = 3 * z1;
     int i = nextInt(z3 + window - 2, gen);
     int j = nextInt(window, gen);
     int k = nextInt(window - 1, gen);
@@ -763,20 +823,6 @@ public final class RandomIndexer {
       SortingNetwork.sort(result, 0, 1, 2);
     }
     return result;
-  }
-
-  private static void setAndAdjustWindowedPair(
-      int[] result, int i, int j, final int z1, final int z2) {
-    if (i < z2) {
-      int x = i & 1;
-      result[x] = i >> 1;
-      result[x ^ 1] = result[x] + 1 + j;
-    } else {
-      i -= z1;
-      j += z1;
-      result[0] = i >= j ? i + 1 : i;
-      result[1] = j;
-    }
   }
 
   private static int iAdjustmentWindowedTriple(int i, int lower, int higher) {